+++ R/Incremental.R
@@ -1,4 +1,4 @@
-# Copyright 2024 Observational Health Data Sciences and Informatics
+# Copyright 2025 Observational Health Data Sciences and Informatics
 #
 # This file is part of CohortGenerator
 #
@@ -178,7 +178,8 @@ recordTasksDone <- function(..., checksum, recordKeepingFile, incremental = TRUE
 #'
 #' @export
 saveIncremental <- function(data, fileName, ...) {
-  if (length(list(...)[[1]]) == 0) {
+  if (length(list(...)) == 0) {
+    warning(paste0("saveIncremental missing key values arguments. Not saving any information to ", fileName))
     return()
   }
   if (file.exists(fileName)) {